requirements.txt
boto3 == 1.34.49
chainlit == 1.0.0
python-dotenv >= 1.0.0
pillow == 10.3.0